Extract CanvasPage from App.js and clean up now-non-nullable types
