+++ src/CanvasPage.css
@@ -1,4 +1,4 @@
-.App {
+.CanvasPage {
   position: absolute;
   top: 0.5rem;
   bottom: 0.5rem;
